Update version and move new training functions to trainingv2
