+++ src/mlops_codex/shared/constants.py
@@ -1,1 +1,1 @@
-CODEX_VERSION = "2.2.8"
+CODEX_VERSION = "2.2.9"
